# docs/swagger.yaml
basePath: /
definitions:
...
  api.CreateCronJobReq:
    properties:
      channel:
        type: string
      feature:
        type: string
      maxRuns:
        type: integer
      model:
        type: string
      name:
        type: string
      prompt:
        type: string
      schedule:
        type: string
      skills:
        items:
          type: string
        type: array
      tags:
        items:
          type: string
        type: array
    required:
    - name
    - prompt
    - schedule
    type: object
...
  api.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  api.UpdateCronJobReq:
    properties:
      channel:
        type: string
      enabled:
        type: boolean
      feature:
        type: string
      maxRuns:
        type: integer
      model:
        type: string
      name:
        type: string
      prompt:
        type: string
      schedule:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 多渠道 AI 聊天机器人框架的 HTTP API
  title: thinkbot API
  version: "1.0"
paths:
...
  /api/bots/{id}/cron:
    get:
      description: 列出指定 Bot 的所有定时任务
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 仅活跃任务
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 定时任务列表
      tags:
      - 定时任务
    post:
      consumes:
      - application/json
      description: 为指定 Bot 创建新的定时任务
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 创建定时任务请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.CreateCronJobReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 创建定时任务
      tags:
      - 定时任务
  /api/bots/{id}/cron/{jobId}:
    delete:
      description: 删除指定的定时任务
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务 ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 删除定时任务
      tags:
      - 定时任务
    get:
      description: 获取指定定时任务的详情
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务 ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 获取定时任务
      tags:
      - 定时任务
    put:
      consumes:
      - application/json
      description: 更新指定定时任务（字段可选）
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务 ID
        in: path
        name: jobId
        required: true
        type: string
      - description: 更新定时任务请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.UpdateCronJobReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 更新定时任务
      tags:
      - 定时任务
  /api/bots/{id}/cron/{jobId}/pause:
    post:
      description: 暂停指定的定时任务
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务 ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 暂停定时任务
      tags:
      - 定时任务
  /api/bots/{id}/cron/{jobId}/resume:
    post:
      description: 恢复指定的定时任务
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务 ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 恢复定时任务
      tags:
      - 定时任务
  /api/bots/{id}/cron/{jobId}/trigger:
    post:
      description: 手动触发指定的定时任务立即执行
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 任务 ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 触发定时任务
      tags:
      - 定时任务
...
  /api/bots/{id}/dreaming/status:
    get:
      description: 返回指定 Bot 的梦境巩固运行时状态
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 梦境巩固状态
      tags:
      - 梦境巩固
  /api/bots/{id}/dreaming/trigger:
    post:
      description: 手动触发指定 Bot 的梦境巩固流程
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 触发梦境巩固
      tags:
      - 梦境巩固
  /api/bots/{id}/memory:
    get:
      description: 查询指定 Bot 的分层记忆（需要 bot.manage 权限，需开启 dreaming）
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      - description: 记忆层级 (L0/L1/L2/L3)
        in: query
        name: tier
        type: string
      - default: 20
        description: 返回条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 查询记忆
      tags:
      - 记忆
  /api/bots/{id}/memory/stats:
    get:
      description: 返回指定 Bot 的记忆统计信息
      parameters:
      - description: Bot ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - CookieAuth: []
      summary: 记忆统计
      tags:
      - 记忆
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"